fixed constraint issue
