+++ app/src/androidTest/java/com/example/ExampleInstrumentedTest.java
@@ -1,0 +1,32 @@
+package com.example;
+
+import android.content.Context;
+import android.support.test.InstrumentationRegistry;
+import android.support.test.runner.AndroidJUnit4;
+
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import static org.junit.Assert.assertEquals;
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+@RunWith(AndroidJUnit4.class)
+public class ExampleInstrumentedTest {
+    @Test
+    public void useAppContext() {
+        // Context of the app under test.
+        Context appContext = InstrumentationRegistry.getTargetContext();
+        System.out.println("OK");
+        assertEquals("com.example.carpooly", appContext.getPackageName());
+        //assertEquals(1, 1);
+    }
+    @Test
+    public void test1_sample(){
+        assertEquals(4, 2 + 2);
+    }
+}
