--- app/src/test/java/com/example/ExampleUnitTest.java
@@ -1,19 +1,0 @@
-package com.example;
-
-import junit.framework.TestCase;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 4);
-    }
-}
