--- app/src/test/java/com/example/carpooly/Model/RideModelTest.java
@@ -1,12 +1,0 @@
-package com.example.carpooly.Model;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class RideModelTest {
-
-    @Test
-    public void setDriverDisplayName() {
-    }
-}
