--- app/src/test/java/com/example/carpooly/Model/UsersSearchTest.java
@@ -1,12 +1,0 @@
-package com.example.carpooly.Model;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class UsersSearchTest {
-
-    @Test
-    public void getDriverName() {
-    }
-}
